+++ mobile/android/android-components/components/service/contile/src/main/java/mozilla/components/service/contile/ContileTopSitesProvider.kt
@@ -5,17 +5,22 @@
 package mozilla.components.service.contile
 
 import android.content.Context
+import android.text.format.DateUtils
 import android.util.AtomicFile
 import androidx.annotation.VisibleForTesting
 import mozilla.components.concept.fetch.Client
+import mozilla.components.concept.fetch.Headers
 import mozilla.components.concept.fetch.Request
+import mozilla.components.concept.fetch.Response
 import mozilla.components.concept.fetch.isSuccess
 import mozilla.components.feature.top.sites.TopSite
 import mozilla.components.feature.top.sites.TopSitesProvider
 import mozilla.components.support.base.log.logger.Logger
 import mozilla.components.support.ktx.android.org.json.asSequence
+import mozilla.components.support.ktx.android.org.json.tryGetLong
 import mozilla.components.support.ktx.util.readAndDeserialize
 import mozilla.components.support.ktx.util.writeString
+import org.json.JSONArray
 import org.json.JSONException
 import org.json.JSONObject
 import java.io.File
@@ -24,7 +29,8 @@ import java.util.Date
 
 internal const val CONTILE_ENDPOINT_URL = "https://contile.services.mozilla.com/v1/tiles"
 internal const val CACHE_FILE_NAME = "mozilla_components_service_contile.json"
-internal const val MINUTE_IN_MS = 60 * 1000
+internal const val CACHE_VALID_FOR_KEY = "valid_for"
+internal const val CACHE_TOP_SITES_KEY = "tiles"
 
 /**
  * Provides access to the Contile services API.
@@ -32,43 +38,44 @@ internal const val MINUTE_IN_MS = 60 * 1000
  * @param context A reference to the application context.
  * @property client [Client] used for interacting with the Contile HTTP API.
  * @property endPointURL The url of the endpoint to fetch from. Defaults to [CONTILE_ENDPOINT_URL].
- * @property maxCacheAgeInMinutes Maximum time (in minutes) the cache should remain valid
- * before a refresh is attempted. Defaults to -1, meaning no cache is being used by default.
+ * @property maxCacheAgeInSeconds Maximum time (in seconds) the cache should remain valid
+ * before a refresh is attempted. Defaults to -1, meaning the max age defined by the server
+ * will be used.
  */
 class ContileTopSitesProvider(
     context: Context,
     private val client: Client,
     private val endPointURL: String = CONTILE_ENDPOINT_URL,
-    private val maxCacheAgeInMinutes: Long = -1,
+    private val maxCacheAgeInSeconds: Long = -1,
 ) : TopSitesProvider {
 
     private val applicationContext = context.applicationContext
     private val logger = Logger("ContileTopSitesProvider")
     private val diskCacheLock = Any()
 
-    // The last modified time of the disk cache.
+    // Current state of the cache.
     @VisibleForTesting
     @Volatile
-    internal var diskCacheLastModified: Long? = null
+    internal var cacheState: CacheState = CacheState(
+        shouldUseServerMaxAge = maxCacheAgeInSeconds == -1L,
+    )
 
     /**
      * Fetches from the top sites [endPointURL] to provide a list of provided top sites.
      * Returns a cached response if [allowCache] is true and the cache is not expired
-     * (@see [maxCacheAgeInMinutes]).
+     * (@see [maxCacheAgeInSeconds]).
      *
      * @param allowCache Whether or not the result may be provided from a previously cached
-     * response. Note that a [maxCacheAgeInMinutes] must be provided in order for the cache to be
-     * active.
+     * response.
      * @throws IOException if the request failed to fetch any top sites.
      */
     @Throws(IOException::class)
     override suspend fun getTopSites(allowCache: Boolean): List<TopSite.Provided> {
         val cachedTopSites = if (allowCache && !isCacheExpired()) {
-            readFromDiskCache()
+            readFromDiskCache()?.topSites
         } else {
             null
         }
-
         if (!cachedTopSites.isNullOrEmpty()) {
             return cachedTopSites
         }
@@ -83,10 +90,10 @@ class ContileTopSitesProvider(
 
     /**
      * Refreshes the cache with the latest top sites response from [endPointURL]
-     * if the cache is active (@see [maxCacheAgeInMinutes]) and expired.
+     * if the cache is expired.
      */
     suspend fun refreshTopSitesIfCacheExpired() {
-        if (maxCacheAgeInMinutes < 0 || !isCacheExpired()) return
+        if (!isCacheExpired()) return
 
         getTopSites(allowCache = false)
     }
@@ -98,12 +105,21 @@ class ContileTopSitesProvider(
             if (response.isSuccess) {
                 val responseBody = response.body.string(Charsets.UTF_8)
 
+                if (response.status == Response.NO_CONTENT) {
+                    // If the response is 204, we should invalidate the cached top sites
+                    cacheState = cacheState.invalidate()
+                    getCacheFile().delete()
+                    return listOf()
+                }
+
                 return try {
-                    JSONObject(responseBody).getTopSites().also {
-                        if (maxCacheAgeInMinutes > 0) {
-                            writeToDiskCache(responseBody)
-                        }
-                    }
+                    val jsonBody = JSONObject(responseBody)
+                    writeToDiskCache(
+                        response.headers.computeValidFor() * DateUtils.SECOND_IN_MILLIS,
+                        jsonBody.getJSONArray(CACHE_TOP_SITES_KEY),
+                    )
+
+                    jsonBody.getTopSites()
                 } catch (e: JSONException) {
                     throw IOException(e)
                 }
@@ -117,51 +133,137 @@ class ContileTopSitesProvider(
     }
 
     @VisibleForTesting
-    internal fun readFromDiskCache(): List<TopSite.Provided>? {
+    internal fun readFromDiskCache(): CachedData? {
         synchronized(diskCacheLock) {
             return getCacheFile().readAndDeserialize {
-                JSONObject(it).getTopSites()
+                JSONObject(it).let { cachedObject ->
+                    CachedData(cachedObject.validFor, cachedObject.getTopSites())
+                }
             }
         }
     }
 
+    /**
+     * Write the validity time and top sites to a file for caching purposes.
+     *
+     * @param validFor Time in milliseconds describing the click validity for the set of top sites.
+     * @param topSites [JSONArray] containing the top sites to be cached.
+     */
     @VisibleForTesting
-    internal fun writeToDiskCache(responseBody: String) {
+    internal fun writeToDiskCache(validFor: Long, topSites: JSONArray) {
+        val cachedData = JSONObject().apply {
+            put(CACHE_VALID_FOR_KEY, validFor)
+            put(CACHE_TOP_SITES_KEY, topSites)
+        }
         synchronized(diskCacheLock) {
             getCacheFile().let {
-                it.writeString { responseBody }
-                diskCacheLastModified = System.currentTimeMillis()
+                it.writeString { cachedData.toString() }
+
+                // Update the cache state to reflect the current status
+                cacheState = cacheState.computeMaxAges(
+                    System.currentTimeMillis(),
+                    maxCacheAgeInSeconds * DateUtils.SECOND_IN_MILLIS,
+                    validFor,
+                )
             }
         }
     }
 
     @VisibleForTesting
-    internal fun isCacheExpired() =
-        getCacheLastModified() < Date().time - maxCacheAgeInMinutes * MINUTE_IN_MS
-
-    @VisibleForTesting
-    internal fun getCacheLastModified(): Long {
-        diskCacheLastModified?.let { return it }
+    internal fun isCacheExpired(): Boolean {
+        cacheState.cacheMaxAge?.let { return Date().time > it }
 
         val file = getBaseCacheFile()
 
-        return if (file.exists()) {
-            file.lastModified().also {
-                diskCacheLastModified = it
+        cacheState =
+            if (file.exists()) {
+                cacheState.computeMaxAges(
+                    file.lastModified(),
+                    maxCacheAgeInSeconds * DateUtils.SECOND_IN_MILLIS,
+                    (readFromDiskCache()?.validFor ?: 0L) * DateUtils.SECOND_IN_MILLIS,
+                )
+            } else {
+                cacheState.invalidate()
             }
-        } else {
-            -1
-        }
+
+        // If cache is invalid, we should also consider it as expired
+        return Date().time > (cacheState.cacheMaxAge ?: -1L)
     }
 
     private fun getCacheFile(): AtomicFile = AtomicFile(getBaseCacheFile())
 
     @VisibleForTesting
     internal fun getBaseCacheFile(): File = File(applicationContext.filesDir, CACHE_FILE_NAME)
+
+    /**
+     * Data stored in the cache file
+     *
+     * @param validFor Time in milliseconds describing the click validity for the set of top sites.
+     * @param topSites List of provided top sites.
+     */
+    internal data class CachedData(
+        val validFor: Long,
+        val topSites: List<TopSite.Provided>,
+    )
+
+    /**
+     * Current state of the cache.
+     *
+     * @param shouldUseServerMaxAge Whether or not [serverCacheMaxAge] should be used instead of
+     * [localCacheMaxAge].
+     * @param isCacheValid Whether or not the current set of cached top sites is still valid.
+     * @param localCacheMaxAge Maximum unix timestamp until the current set of cached top sites
+     * is still valid, specified by the client.
+     * @param serverCacheMaxAge Maximum unix timestamp until the current set of cached top sites
+     * is still valid, specified by the server.
+     */
+    internal data class CacheState(
+        val shouldUseServerMaxAge: Boolean,
+        val isCacheValid: Boolean = true,
+        val localCacheMaxAge: Long? = null,
+        val serverCacheMaxAge: Long? = null,
+    ) {
+        val cacheMaxAge
+            get() = if (isCacheValid) {
+                if (shouldUseServerMaxAge) serverCacheMaxAge else localCacheMaxAge
+            } else {
+                null
+            }
+
+        fun invalidate(): CacheState =
+            this.copy(isCacheValid = false, localCacheMaxAge = null, serverCacheMaxAge = null)
+
+        /**
+         * Update local and server max age values
+         */
+        fun computeMaxAges(lastModified: Long, localMaxAge: Long, serverMaxAge: Long): CacheState =
+            this.copy(
+                isCacheValid = true,
+                localCacheMaxAge = lastModified + localMaxAge,
+                serverCacheMaxAge = lastModified + serverMaxAge,
+            )
+    }
 }
 
+/**
+ * To extract the `valid-for` value for the set of provided top sites, we need to sum up the `max-age`
+ * and `stale-if-error` options from the header. These values can be found in the `cache-control` header,
+ * formatted as `max-age=$value` and `stale-if-error=$value`.
+ */
+internal fun Headers.computeValidFor(): Long =
+    getAll("cache-control").sumOf {
+        val valueList = it.split("=")
+            .map { item -> item.trim() }
+
+        if (valueList.size == 2 && valueList[0] in listOf("max-age", "stale-if-error")) {
+            valueList[1].toLong()
+        } else {
+            0L
+        }
+    }
+
 internal fun JSONObject.getTopSites(): List<TopSite.Provided> =
-    getJSONArray("tiles")
+    getJSONArray(CACHE_TOP_SITES_KEY)
         .asSequence { i -> getJSONObject(i) }
         .mapNotNull { it.toTopSite() }
         .toList()
@@ -181,3 +283,6 @@ private fun JSONObject.toTopSite(): TopSite.Provided? {
         null
     }
 }
+
+internal val JSONObject.validFor: Long
+    get() = this.tryGetLong(CACHE_VALID_FOR_KEY) ?: 0L